+++ packages/dialect-react-ui/components/Chat/screens/CreateThreadPage/CreateThread.tsx
@@ -377,7 +377,10 @@ export default function CreateThread({
     setIsTyping(true);
   };
 
-  const findAddress = async (connection: Connection, addressString: string) => {
+  const findAddress = async (
+    connection: Connection | undefined,
+    addressString: string
+  ) => {
     if (!connection) {
       // TODO: set connection error
       return;
@@ -426,11 +429,12 @@ export default function CreateThread({
     setIsTyping(false);
   };
 
+  // TODO: Fix Eslint
   const findAddressDebounced = useCallback(debounce(findAddress, 500), []);
 
   useEffect(() => {
     // When input address changes, we debounce
-    findAddressDebounced(connection, address);
+    findAddressDebounced(connection, address ?? '');
   }, [findAddressDebounced, connection, address]);
 
   useEffect(() => {